feat: move 명령어 구현
