+++ src/main/resources/templates/index.html
@@ -5,6 +5,9 @@
     <title>로또</title>
 </head>
 <body>
-Hello World!!
+    <form method="post" action="/move">
+        <input type="text" name="position" placeholder="좌표를 입력해 주세요.">
+        <button type="submit" value="제출">제출</button>
+    </form>
 </body>
 </html>
